feat: WI-0479 split payslips compare helpers
