+++ ROADMAP.md
@@ -743,3 +743,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0476 admin non-payroll workspace line-budget guard (`/admin/notices`, `/admin/benefits`, `/admin/recruitment` line-budget <=300 회귀 가드 + phase-loop 금지 패턴 회귀 차단 테스트 추가)
 - WI-0477 korean copy residual sweep for payroll preset preview (`PayrollKrIncomeSplitPresetPayloadPreviewPanel.tsx` ko copy block 독립화 + preset mode 생략 라벨 locale wiring + payroll preset preview 한글 회귀 테스트 추가)
 - WI-0478 payslips page-view filter/list panel decomposition (`/employee/payslips`의 page-view에서 필터/개발도구/근태요약 패널과 명세서 목록 패널 분리 + `page-view-types.ts` 타입 추출 + `page-view.tsx` 461->209 line-budget 회복 + regression test)
+- WI-0479 payslips helper budget split (`page-helpers.ts`의 비교/인사이트 계산 블록을 `page-compare-helpers.ts`로 분리 + 기존 호출부는 re-export로 호환 유지 + legacy WI-0318/0319/0396/0398 회귀 테스트를 분해 구조에 맞게 보정 + `page-helpers.ts` 394->230)
